ml/functions/movies-knn-train/requirements.txt
functions-framework
pandas
numpy
joblib
gcsfs
scikit-learn
google-cloud-storage
google-cloud-aiplatform
pyarrow